+++ sctoolbox/tools/__init__.py
@@ -1,2 +1,3 @@
+"""sctoolbox tool functions."""
 from sctoolbox._modules import import_submodules
 function_dict = import_submodules(globals())