+++ sctoolbox/utils/decorator.py
@@ -1,14 +1,18 @@
+"""Decorators and related functions."""
+
 import anndata
 import functools
 import pandas as pd
 import matplotlib
 
+from typing import Callable
+
 import sctoolbox.utils.general as utils
 
 
-def log_anndata(func):
+def log_anndata(func) -> Callable:
     """
-    Decorator to log function call inside adata.
+    Decorate function to log adata inside function call.
 
     Parameters
     ----------
@@ -17,9 +21,10 @@ def log_anndata(func):
 
     Returns
     -------
-    function :
+    Callable :
         Decorated function
     """
+
     # TODO store datatypes not supported by scanpy.write as string representation (repr())
 
     @functools.wraps(func)  # preserve information of the decorated func
@@ -70,7 +75,7 @@ def log_anndata(func):
     return wrapper
 
 
-def get_parameter_table(adata):
+def get_parameter_table(adata) -> pd.DataFrame:
     """
     Get a table of all function calls with their parameters from the adata.uns["sctoolbox"] dictionary.
 
@@ -83,6 +88,11 @@ def get_parameter_table(adata):
     -------
     pd.DataFrame
         Table with all function calls and their parameters.
+
+    Raises
+    ------
+    ValueError
+        If no logs are found.
     """
 
     if "sctoolbox" not in adata.uns.keys() or "log" not in adata.uns["sctoolbox"].keys():
@@ -108,10 +118,16 @@ def get_parameter_table(adata):
     return complete_table
 
 
-def debug_func_log(func):
-    """ Decorator to print function call with arguments and keyword arguments.
+def debug_func_log(func) -> None:
+    """
+    Decorate function to print function call with arguments and keyword arguments.
 
     In progress.
+
+    Parameters
+    ----------
+    func : function
+        Function to decorate.
     """
 
     @functools.wraps(func)