+++ sctoolbox/utils/multiprocessing.py
@@ -1,8 +1,16 @@
+"""Functions related to multiprocessing."""
+
 import time
 import sctoolbox.utils as utils
 
+# type hint imports
+from typing import TYPE_CHECKING
 
-def get_pbar(total, description):
+if TYPE_CHECKING:
+    import tqdm
+
+
+def get_pbar(total, description) -> "tqdm.tqdm":
     """
     Get a progress bar depending on whether the user is using a notebook or not.
 
@@ -15,7 +23,7 @@ def get_pbar(total, description):
 
     Returns
     -------
-    tqdm
+    tqdm.tqdm
         A progress bar object.
     """
 
@@ -28,7 +36,7 @@ def get_pbar(total, description):
     return pbar
 
 
-def monitor_jobs(jobs, description="Progress"):
+def monitor_jobs(jobs, description="Progress") -> None:
     """
     Monitor the status of jobs submitted to a pool.