+++ tests/test_atac.py
@@ -1,3 +1,5 @@
+"""Test atac functions."""
+
 import pytest
 import sctoolbox.atac
 import os
@@ -11,7 +13,7 @@ import sctoolbox.plotting as pl
 
 @pytest.fixture
 def adata():
-    """ Fixture for an AnnData object. """
+    """Fixture for an AnnData object."""
     adata = sc.read_h5ad(os.path.join(os.path.dirname(__file__), 'data', 'atac', 'mm10_atac.h5ad'))
     return adata
 
@@ -19,6 +21,7 @@ def adata():
 # adapted from muon package
 @pytest.fixture
 def tfidf_x():
+    """Create anndata with random expression."""
     np.random.seed(2020)
     x = np.abs(np.random.normal(size=(4, 5)))
     adata_X = ad.AnnData(x)
@@ -27,12 +30,14 @@ def tfidf_x():
 
 # adapted from muon package
 def test_tfidf(tfidf_x):
+    """Test tfidt success."""
     sctoolbox.atac.tfidf(tfidf_x, log_tf=True, log_idf=True)
     assert str("%.3f" % tfidf_x.X[0, 0]) == "4.659"
     assert str("%.3f" % tfidf_x.X[3, 0]) == "4.770"
 
 
 def test_lsi(adata):
+    """Test lsi success."""
     sctoolbox.atac.tfidf(adata)
     sctoolbox.atac.lsi(adata)
     assert "X_lsi" in adata.obsm and "lsi" in adata.uns and "LSI" in adata.varm
@@ -40,6 +45,7 @@ def test_lsi(adata):
 
 @pytest.mark.parametrize("method", ["tfidf", "total"])
 def test_atac_norm(adata, method):
+    """Test atac_norm success."""
     adata_norm = sctoolbox.atac.atac_norm(adata, method=method)[method]  # return from function is a dict
 
     if method == "tfidf":
@@ -49,6 +55,7 @@ def test_atac_norm(adata, method):
 
 
 def test_write_TOBIAS_config():
+    """Test write_TOBIAS_config success."""
 
     sctoolbox.atac.write_TOBIAS_config("tobias.yml", bams=["bam1.bam", "bam2.bam"])
     yml = yaml.full_load(open("tobias.yml"))
@@ -57,7 +64,7 @@ def test_write_TOBIAS_config():
 
 
 def test_add_insertsize_fragments(adata):
-    """ Test if add_insertsize adds information from a fragmentsfile """
+    """Test if add_insertsize adds information from a fragmentsfile."""
 
     adata = adata.copy()
     fragments = os.path.join(os.path.dirname(__file__), 'data', 'atac', 'mm10_atac_fragments.bed')
@@ -68,7 +75,7 @@ def test_add_insertsize_fragments(adata):
 
 
 def test_add_insertsize_bam(adata):
-    """ Test if add_insertsize adds information from a bamfile """
+    """Test if add_insertsize adds information from a bamfile."""
 
     adata = adata.copy()
     bam = os.path.join(os.path.dirname(__file__), 'data', 'atac', 'mm10_atac.bam')
@@ -79,7 +86,7 @@ def test_add_insertsize_bam(adata):
 
 
 def test_insertsize_plotting(adata):
-    """ Test if insertsize plotting works """
+    """Test if insertsize plotting works."""
 
     adata = adata.copy()
     fragments = os.path.join(os.path.dirname(__file__), 'data', 'atac', 'mm10_atac_fragments.bed')