+++ CHANGES.rst
@@ -11,8 +11,15 @@
 - Added check of column validity to tools.marker_genes.run_DESeq2() (#134)
 - Increase test coverage for plotting functions (#126)
 - Apply fixes to bugs found by increasing the test coverage.
+- Added type hinting to functions.
+- Revised doc-strings.
+- run_rank_genes() auto converts groupby column to type 'category' (#137)
+- Fix parameter for gene/cell filtering (#136)
+- Add Check to _filter_object() if column contains only boolean (#110)
+- Add support of matrx and numpy.ndarray type of adata.X for predict_sex (#111)
+- Add method to get pd.DataFrame columns with list of regex (#90)
 
-Changes to ATAC notebooks (0.1.1.2 - 05-06-2023):
+## Changes to ATAC notebooks (0.1.1.2 - 05-06-2023):   
 - Fixed assembling atac notebook 01
 - Fixed get_atac_thresholds_wrapper and renamed it to get_thresholds_wrapper
 - Added custome cwt implementation